Modified Apptainer configuration
Made a few updates in lieu of the updates to the planner submission rules.

Removed Tarski variant from the Satisificing track. Hence, the repo is now associated with 3 different runs instead of 4.

The three configurations of the planner are

Agile 1. The standard configuration of ApxNovelty planner with FD parser and grounder for the Agile track.
Agile 2. A variant of ApxNovelty search that uses Tarski grounder. We expect that the reduced time-taken to ground a PDDL instance would have a positive impact in the Agile track which has a time-limit of 300 seconds.
Satisficing 1. The anytime configuration of the planner. It uses ApxNovelty search to find the first satisficing plan. After which, it switches to lazy wastar with a upper bound on the plan cost, and continues to optimize the solutions until timeout.